requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.3
django-elasticsearch-dsl==7.3
djangorestframework==3.14.0
drf-yasg==1.21.6
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.0
uritemplate==4.1.1
urllib3==1.26.16

django-filter~=23.2